requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
yt-dlp==2023.11.16
openai-whisper==20231117
ffmpeg-python==0.2.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2 